cleanup synchronisation
interim solution of synchronising on the manager factory is not
necessary anymore. It is generally known to be threadsafe (or at least
it should be if it isn't currently), so synchronisation is only
necessary on OWLOntologyManager instances when they are accessed through
public APIs. Internally we do not synchronize on OWLOntologyManager
objects that are passed into internal methods.