work on ensuring that the schema management graph is consistent for tests
the internal changes highlighted instances where the schema management
graph was not being maintained by tests, as they previously did not
actually rely on it for unit tests

with the advent of compulsory caching based solely on the schema
management graph, these tests started to massively fail, and there isn't
a single solution as they all implemented schema loading differently.
this commit and others coming up will try to fix that.